+++ test/lib/_database_test.js
@@ -1,3 +1,5 @@
+var Ddl = require("../..")
+
 exports.mustPassSimpleTable = function(withSql) {
   withSql(
     'CREATE TABLE "test" (' +
@@ -8,43 +10,47 @@ exports.mustPassSimpleTable = function(withSql) {
     ')'
   )
 
+  it("must return an instance of Ddl", function() {
+    this.ddl.must.be.an.instanceof(Ddl)
+  })
+
   it("must return all column names", function() {
-    this.attrs.must.have.keys(["id", "name", "age", "notes"])
+    this.ddl.must.have.keys(["id", "name", "age", "notes"])
   })
 
   it("must return null statuses", function() {
-    this.attrs.id.null.must.be.false()
-    this.attrs.name.null.must.be.false()
-    this.attrs.age.null.must.be.true()
-    this.attrs.notes.null.must.be.true()
+    this.ddl.id.null.must.be.false()
+    this.ddl.name.null.must.be.false()
+    this.ddl.age.null.must.be.true()
+    this.ddl.notes.null.must.be.true()
   })
 
   it("must not have duplicate names", function() {
-    this.attrs.id.must.not.have.property("name")
-    this.attrs.name.must.not.have.property("name")
-    this.attrs.age.must.not.have.property("name")
-    this.attrs.notes.must.not.have.property("name")
+    this.ddl.id.must.not.have.property("name")
+    this.ddl.name.must.not.have.property("name")
+    this.ddl.age.must.not.have.property("name")
+    this.ddl.notes.must.not.have.property("name")
   })
 
   it("must return types", function() {
-    this.attrs.id.type.must.equal(Number)
-    this.attrs.name.type.must.equal(String)
-    this.attrs.age.type.must.equal(Number)
-    this.attrs.notes.type.must.equal(String)
+    this.ddl.id.type.must.equal(Number)
+    this.ddl.name.type.must.equal(String)
+    this.ddl.age.type.must.equal(Number)
+    this.ddl.notes.type.must.equal(String)
   })
 
   it("must return defaults", function() {
-    this.attrs.id.must.have.property("default", null)
-    this.attrs.name.must.have.property("default", null)
-    this.attrs.age.must.have.property("default", 18)
-    this.attrs.notes.must.have.property("default", "")
+    this.ddl.id.must.have.property("default", null)
+    this.ddl.name.must.have.property("default", null)
+    this.ddl.age.must.have.property("default", 18)
+    this.ddl.notes.must.have.property("default", "")
   })
 
   it("must return limits", function() {
-    this.attrs.id.must.have.property("limit", null)
-    this.attrs.name.must.have.property("limit", 255)
-    this.attrs.age.must.have.property("limit", null)
-    this.attrs.notes.must.have.property("limit", null)
+    this.ddl.id.must.have.property("limit", null)
+    this.ddl.name.must.have.property("limit", 255)
+    this.ddl.age.must.have.property("limit", null)
+    this.ddl.notes.must.have.property("limit", null)
   })
 }
 
@@ -53,7 +59,7 @@ exports.mustPassDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" INTEGER DEFAULT NULL)')
 
     it("must be set to null", function() {
-      this.attrs.foo.must.have.property("default", null)
+      this.ddl.foo.must.have.property("default", null)
     })
   })
 
@@ -61,7 +67,7 @@ exports.mustPassDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" INTEGER DEFAULT null)')
 
     it("must be set to null", function() {
-      this.attrs.foo.must.have.property("default", null)
+      this.ddl.foo.must.have.property("default", null)
     })
   })
 
@@ -69,7 +75,7 @@ exports.mustPassDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" TEXT DEFAULT (1 + 2))')
 
     it("must be set to undefined", function() {
-      this.attrs.foo.must.have.property("default", undefined)
+      this.ddl.foo.must.have.property("default", undefined)
     })
   })
 }
@@ -79,7 +85,7 @@ exports.mustPassRealDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" REAL DEFAULT 42.69)')
 
     it("must be set to 42.69", function() {
-      this.attrs.foo.must.have.property("default", 42.69)
+      this.ddl.foo.must.have.property("default", 42.69)
     })
   })
 
@@ -87,7 +93,7 @@ exports.mustPassRealDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" REAL DEFAULT 42e3)')
 
     it("must be set to 42000", function() {
-      this.attrs.foo.must.have.property("default", 42000)
+      this.ddl.foo.must.have.property("default", 42000)
     })
   })
 
@@ -95,7 +101,7 @@ exports.mustPassRealDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" REAL DEFAULT 42.511e3)')
 
     it("must be set to 42511", function() {
-      this.attrs.foo.must.have.property("default", 42511)
+      this.ddl.foo.must.have.property("default", 42511)
     })
   })
 
@@ -103,7 +109,7 @@ exports.mustPassRealDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" REAL DEFAULT 42.)')
 
     it("must be set to 42", function() {
-      this.attrs.foo.must.have.property("default", 42)
+      this.ddl.foo.must.have.property("default", 42)
     })
   })
 
@@ -111,7 +117,7 @@ exports.mustPassRealDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" REAL DEFAULT .42)')
 
     it("must be set to 0.42", function() {
-      this.attrs.foo.must.have.property("default", .42)
+      this.ddl.foo.must.have.property("default", 0.42)
     })
   })
 
@@ -119,7 +125,7 @@ exports.mustPassRealDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" REAL DEFAULT 42.e3)')
 
     it("must be set to 42000", function() {
-      this.attrs.foo.must.have.property("default", 42000)
+      this.ddl.foo.must.have.property("default", 42000)
     })
   })
 
@@ -127,7 +133,7 @@ exports.mustPassRealDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" REAL DEFAULT -42.e3)')
 
     it("must be set to -42000", function() {
-      this.attrs.foo.must.have.property("default", -42000)
+      this.ddl.foo.must.have.property("default", -42000)
     })
   })
 }
@@ -137,7 +143,7 @@ exports.mustPassTextDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" TEXT DEFAULT \'a b c\')')
 
     it("must be set", function() {
-      this.attrs.foo.default.must.equal("a b c")
+      this.ddl.foo.default.must.equal("a b c")
     })
   })
 
@@ -145,7 +151,7 @@ exports.mustPassTextDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" TEXT DEFAULT \'"\')')
 
     it("must be set to \"", function() {
-      this.attrs.foo.default.must.equal("\"")
+      this.ddl.foo.default.must.equal("\"")
     })
   })
 
@@ -153,7 +159,7 @@ exports.mustPassTextDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" TEXT DEFAULT \'\'\'\')')
 
     it("must be set to '", function() {
-      this.attrs.foo.default.must.equal("'")
+      this.ddl.foo.default.must.equal("'")
     })
   })
 
@@ -161,7 +167,7 @@ exports.mustPassTextDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" TEXT DEFAULT \'""\')')
 
     it("must be set to \"\"", function() {
-      this.attrs.foo.default.must.equal("\"\"")
+      this.ddl.foo.default.must.equal("\"\"")
     })
   })
 }
@@ -171,7 +177,7 @@ exports.mustPassBooleanDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" BOOLEAN DEFAULT true)')
 
     it("must be set to true", function() {
-      this.attrs.foo.must.have.property("default", true)
+      this.ddl.foo.must.have.property("default", true)
     })
   })
 
@@ -179,7 +185,7 @@ exports.mustPassBooleanDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" BOOLEAN DEFAULT false)')
 
     it("must be set to false", function() {
-      this.attrs.foo.must.have.property("default", false)
+      this.ddl.foo.must.have.property("default", false)
     })
   })
 
@@ -187,7 +193,7 @@ exports.mustPassBooleanDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" BOOLEAN DEFAULT \'t\')')
 
     it("must be set to true", function() {
-      this.attrs.foo.must.have.property("default", true)
+      this.ddl.foo.must.have.property("default", true)
     })
   })
 
@@ -195,7 +201,7 @@ exports.mustPassBooleanDefault = function(withSql) {
     withSql('CREATE TABLE "test" ("foo" BOOLEAN DEFAULT \'f\')')
 
     it("must be set to false", function() {
-      this.attrs.foo.must.have.property("default", false)
+      this.ddl.foo.must.have.property("default", false)
     })
   })
 }