+++ test/lib/postgresql_test.js
@@ -1,22 +1,17 @@
 var _ = require("underscore")
 var Pg = require("pg")
+var ddl = require("../..").postgresql
 
-Pg.defaults.host = "/tmp"
-Pg.defaults.database = "assertions_test"
-
-var db = new Pg.Client
-before(function(done) { db.connect(done) })
+var db = new Pg.Client({host: "/tmp", database: "ddl_test"})
+before(db.connect.bind(db))
 after(db.end.bind(db))
 
-var Shared = require("./shared")
-var attributes = require("..").postgresql
-
-describe("PostgreSQL", function() {
+describe("Ddl.postgresql", function() {
   beforeEach(function(done) { db.query("BEGIN", done) })
   afterEach(function(done) { db.query("ROLLBACK", done) })
 
   describe("given a simple table", function() {
-    Shared.mustPassSimpleTable(withSql)
+    require("./_database_test").mustPassSimpleTable(withSql)
   })
 
   describe("type", function() {
@@ -44,7 +39,7 @@ describe("PostgreSQL", function() {
         withSql('CREATE TABLE "test" ("foo" ' + type + ')')
 
         it("must be set to " + klass.name, function() {
-          this.attrs.foo.type.must.equal(klass)
+          this.ddl.foo.type.must.equal(klass)
         })
       })
     })
@@ -53,24 +48,24 @@ describe("PostgreSQL", function() {
       withSql('CREATE TABLE "test" ("foo" Date)')
 
       it("must be set properly", function() {
-        this.attrs.foo.type.must.equal(Date)
+        this.ddl.foo.type.must.equal(Date)
       })
     })
   })
 
   describe("default", function() {
-    Shared.mustPassDefault(withSql)
+    require("./_database_test").mustPassDefault(withSql)
 
     describe("given autoincrement", function() {
       withSql('CREATE TABLE "test" ("foo" SERIAL)')
 
       it("must be set to undefined", function() {
-        this.attrs.foo.must.have.property("default", undefined)
+        this.ddl.foo.must.have.property("default", undefined)
       })
     })
 
     describe("of TEXT column", function() {
-      Shared.mustPassTextDefault(withSql)
+      require("./_database_test").mustPassTextDefault(withSql)
 
       describe("given 'unknown'::character varying", function() {
         withSql('CREATE TABLE "test" ' +
@@ -79,23 +74,23 @@ describe("PostgreSQL", function() {
           ')')
 
         it("must be set to \"unknown\"", function() {
-          this.attrs.foo.must.have.property("default", "unknown")
+          this.ddl.foo.must.have.property("default", "unknown")
         })
       })
     })
 
     describe("of REAL column", function() {
-      Shared.mustPassRealDefault(withSql)
+      require("./_database_test").mustPassRealDefault(withSql)
     })
 
     describe("of BOOLEAN column", function() {
-      Shared.mustPassBooleanDefault(withSql)
+      require("./_database_test").mustPassBooleanDefault(withSql)
 
       describe("given 1::boolean", function() {
         withSql('CREATE TABLE "test" ("foo" BOOLEAN DEFAULT 1::boolean)')
 
         it("must be set to true", function() {
-          this.attrs.foo.must.have.property("default", true)
+          this.ddl.foo.must.have.property("default", true)
         })
       })
 
@@ -103,7 +98,7 @@ describe("PostgreSQL", function() {
         withSql('CREATE TABLE "test" ("foo" BOOLEAN DEFAULT \'t\'::boolean)')
 
         it("must be set to true", function() {
-          this.attrs.foo.must.have.property("default", true)
+          this.ddl.foo.must.have.property("default", true)
         })
       })
 
@@ -111,7 +106,7 @@ describe("PostgreSQL", function() {
         withSql('CREATE TABLE "test" ("foo" BOOLEAN DEFAULT 0::boolean)')
 
         it("must be set to false", function() {
-          this.attrs.foo.must.have.property("default", false)
+          this.ddl.foo.must.have.property("default", false)
         })
       })
 
@@ -119,7 +114,7 @@ describe("PostgreSQL", function() {
         withSql('CREATE TABLE "test" ("foo" BOOLEAN DEFAULT \'f\'::boolean)')
 
         it("must be set to false", function() {
-          this.attrs.foo.must.have.property("default", false)
+          this.ddl.foo.must.have.property("default", false)
         })
       })
 
@@ -127,7 +122,7 @@ describe("PostgreSQL", function() {
         withSql('CREATE TABLE "test" ("foo" BOOLEAN DEFAULT 42::boolean)')
 
         it("must be set to undefined", function() {
-          this.attrs.foo.must.have.property("default", undefined)
+          this.ddl.foo.must.have.property("default", undefined)
         })
       })
     })
@@ -139,7 +134,7 @@ describe("PostgreSQL", function() {
         )
 
         it("must be set to \"00:00:00\"", function() {
-          this.attrs.foo.must.have.property("default", "00:00:00")
+          this.ddl.foo.must.have.property("default", "00:00:00")
         })
       })
     })
@@ -147,15 +142,15 @@ describe("PostgreSQL", function() {
 
   function withSql(sql, fn) {
     beforeEach(function(done) { db.query(sql, done) })
-    beforeEach(withAttrs(function(attrs) { this.attrs = attrs }))
+    beforeEach(withAttrs(function(ddl) { this.ddl = ddl }))
   }
 
   function withAttrs(fn) {
     return function(done) {
       var self = this
-      attributes(db, "test", function(err, attrs) {
+      ddl(db, "test", function(err, ddl) {
         if (err) return done(err)
-        fn.call(self, attrs)
+        fn.call(self, ddl)
         done()
       })
     }